adding tv show details
